+++ samples/TypeShape.Samples/expr.fs
@@ -31,25 +31,25 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     let meaningUntyped (e : Expr) =
         TypeShape.Create(e.Type).Accept {
-            new ITypeShapeVisitor<CompiledExpr<obj>> with
+            new ITypeVisitor<CompiledExpr<obj>> with
                 member __.Visit<'t>() =
                     let ce = meaning<'t>(cast e)
                     fun env -> ce env :> obj
         }
 
-    let mkMemberWriter (s : IShapeWriteMember<'T>) (e : Expr) =
-        s.Accept { new IWriteMemberVisitor<'T, CompiledExpr<'T -> 'T>> with
-            member __.Visit(s : ShapeWriteMember<'T, 'Elem>) =
+    let mkMemberWriter (s : IShapeMember<'T>) (e : Expr) =
+        s.Accept { new IMemberVisitor<'T, CompiledExpr<'T -> 'T>> with
+            member __.Visit(s : ShapeMember<'T, 'Elem>) =
                 let cElem = meaning<'Elem> (cast e)
-                fun env t -> let e = cElem env in s.Inject t e }
+                fun env t -> let e = cElem env in s.Set t e }
 
-    let mkProjection (s : IShapeMember) (value : Expr) =
+    let mkProjection (s : IShapeReadOnlyMember) (value : Expr) =
         TypeShape.Create(value.Type).Accept { 
-            new ITypeShapeVisitor<CompiledExpr<'T>> with
+            new ITypeVisitor<CompiledExpr<'T>> with
                 member __.Visit<'Value> () =
-                    let s = s :?> ShapeMember<'Value, 'T>
+                    let s = s :?> ReadOnlyMember<'Value, 'T>
                     let cvalue = meaning<'Value> (cast value)
-                    fun env -> let v = cvalue env in s.Project v }
+                    fun env -> let v = cvalue env in s.Get v }
 
     match expr with
     | Value(:? 'T as t, _)
@@ -58,7 +58,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     | Application(func, arg) ->
         let argShape = TypeShape.Create arg.Type
-        argShape.Accept { new ITypeShapeVisitor<CompiledExpr<'T>> with
+        argShape.Accept { new ITypeVisitor<CompiledExpr<'T>> with
             member __.Visit<'Arg>() =
                 let cfunc = meaning<'Arg -> 'T> (cast func)
                 let carg = meaning<'Arg> (cast arg)
@@ -76,7 +76,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     | Let(var, bind, cont) ->
         let vShape = TypeShape.Create var.Type
-        vShape.Accept { new ITypeShapeVisitor<CompiledExpr<'T>> with
+        vShape.Accept { new ITypeVisitor<CompiledExpr<'T>> with
             member __.Visit<'Var>() = 
                 let cbind = meaning<'Var> (cast bind)
                 let ccont = meaning<'T> (cast cont)
@@ -226,7 +226,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
     | LetRecursive(recBindings, cont) ->
         let mkRecBinding (var : Var, body : Expr) =
             let s = TypeShape.Create var.Type
-            s.Accept { new ITypeShapeVisitor<Var * (Environment -> unit)> with
+            s.Accept { new ITypeVisitor<Var * (Environment -> unit)> with
                 member __.Visit<'Body>() =
                     let cbody = meaning<'Body> (cast body)
                     var, fun env -> env.UpdateVar(var, cbody env) }
@@ -246,7 +246,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     | VarSet(var, expr) ->
         TypeShape.Create(expr.Type).Accept {
-          new ITypeShapeVisitor<CompiledExpr<'T>> with
+          new ITypeVisitor<CompiledExpr<'T>> with
             member __.Visit<'t>() =
                 let cexpr = meaning<'t>(cast expr)
                 EQ(fun env -> env.UpdateVar(var, cexpr env))
@@ -254,7 +254,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     | Coerce(target, _) ->
         TypeShape.Create(target.Type).Accept {
-          new ITypeShapeVisitor<CompiledExpr<'T>> with
+          new ITypeVisitor<CompiledExpr<'T>> with
             member __.Visit<'t>() =
                 let cexpr = meaning<'t>(cast expr)
                 fun env -> cexpr env |> unbox<'T>
@@ -262,11 +262,11 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
 
     | TypeTest(target, ty) ->
         TypeShape.Create(target.Type).Accept {
-          new ITypeShapeVisitor<CompiledExpr<'T>> with
+          new ITypeVisitor<CompiledExpr<'T>> with
             member __.Visit<'t>() =
               let ctarget = meaning<'t>(cast expr)
               TypeShape.Create(ty).Accept {
-                new ITypeShapeVisitor<CompiledExpr<'T>> with
+                new ITypeVisitor<CompiledExpr<'T>> with
                   member __.Visit<'s>() =
                     EQ(fun env -> box (ctarget env) :? 's)
               }
@@ -275,7 +275,7 @@ let rec meaning<'T> (expr : Expr<'T>) : CompiledExpr<'T> =
     | QuoteRaw q -> EQ(fun env -> env.Splice q)
     | QuoteTyped q ->
         TypeShape.Create(q.Type).Accept {
-            new ITypeShapeVisitor<CompiledExpr<'T>> with
+            new ITypeVisitor<CompiledExpr<'T>> with
                 member __.Visit<'t> () = // 'T = Expr<'t>
                     EQ(fun env -> Expr.Cast<'t> (env.Splice q))
         }